Log-Sum-Exp trick on log-likelihood
The power of this trick is seen primarily in the likelihood, because it works out to where every exponent is shifted by the maximum, and you never actually have to take the exponent of the maximum. This is particularly important during the first gradient update, because it might take a gradient step that makes it impossible to compute the likelihood without this trick. I did the same thing for the derivatives with respect to gamma and d1, d2, but it is likely not as useful in that case (same reason it is not useful for chi). Btw, the reason I need the likelihood is the line search.